backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
fpdf2==2.8.3
supabase==2.15.1
